Refactor add methods
